requirements.txt
altgraph==0.17.4
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromedriver-autoinstaller-max==1.0.6
coloredlogs==15.0.1
cryptography==45.0.6
ddddocr==1.5.6
Deprecated==1.2.18
flatbuffers==25.2.10
frozendict==2.4.6
h11==0.16.0
humanfriendly==10.0
idna==3.10
looseversion==1.3.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multitasking==0.0.11
nodriver==0.47.0
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
peewee==3.18.1
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
playsound==1.2.2
protobuf==6.32.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
selenium==4.35.0
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tornado==6.5.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
windows-capture==1.4.4
wrapt==1.17.3
wsproto==1.2.0
yfinance==0.2.54
